net/tcp: Fix 64K data duplication from uint16_t truncation in net_incr32
net_incr32() accepted a uint16_t increment, but callers pass
io_pktlen which is 32-bit. When merged OFO segments exceed 65535
bytes, the increment is silently truncated, causing rcvseq to
under-advance vs readahead growth. This delivers 65536 bytes of
duplicate data to the application.

Fix by widening net_incr32() to uint32_t, and fix related
uint16_t truncation in tcp_datahandler(), tcp_data_event(),
tcp_newdata(), and net_iob_concat().

+++ net/tcp/tcp.h
@@ -1443,7 +1443,7 @@ uint32_t tcp_callback(FAR struct net_driver_s *dev,
  *
  ****************************************************************************/
 
-uint16_t tcp_datahandler(FAR struct net_driver_s *dev,
+uint32_t tcp_datahandler(FAR struct net_driver_s *dev,
                          FAR struct tcp_conn_s *conn,
                          uint16_t offset);
