+++ net/utils/net_iob_concat.c
@@ -49,7 +49,7 @@
  *
  ****************************************************************************/
 
-uint16_t net_iob_concat(FAR struct iob_s **iob1, FAR struct iob_s **iob2)
+uint32_t net_iob_concat(FAR struct iob_s **iob1, FAR struct iob_s **iob2)
 {
   if (*iob1 == NULL)
     {
